requirements.txt
# FinanzBro - Aktienportfolio Dashboard
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pandas==2.2.3
jinja2==3.1.5
pydantic==2.10.4

# Finanzdaten
yfinance==0.2.48
websockets==14.1

# Scheduling
apscheduler==3.10.4

# AI (Gemini Research, optional)
google-genai>=1.0.0

# Charts (served via CDN, but listed for reference)
# chart.js - loaded via CDN in frontend

# Testing
pytest>=8.0
pytest-asyncio>=0.23
